requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pydantic
pydantic-settings
cryptography
pyjwt[crypto]